refactor(scaffold): split init.ts into focused modules + a template-data module
init.ts had grown to 611 lines and 37 revisions — the all-time churn champion —
because it fused four concerns: ~250 lines of embedded template strings (data) plus
three command mechanisms (init / add channel / add skill). Every scaffold feature
edited this one file, and a template tweak collided with a logic change.

Split along those seams (behavior unchanged; pure move + import rewiring):
- scaffold-templates.ts — all file content init/add write (data), incl. agentsMd,
  packageJson, toPackageName, and the channel/tool templates.
- add-channel.ts — `add <channel>`: CHANNEL_SCAFFOLDS + scaffoldChannel/appendChannelEnv/
  channelSetup/channelExists/assertChannelReady.
- vendor-skill.ts — `add skill`: vendorSkill + source resolution (giget/local/bare-name).
- init.ts — now just scaffoldWorkspace + the shared `exists` helper.

cli.ts imports from the three modules; index.ts's scaffoldWorkspace/ScaffoldResult
re-export is unchanged (both stayed in init.ts); init.test.ts points vendorSkill at
its new module. Templates stay TS string literals (no build change) — converting them
to real files to drop the backtick/backslash escaping needs a bundler loader and is
deferred.
