+++ core/src/engines/pi/init.ts
@@ -7,117 +7,24 @@
  * Scope: init is best-effort atomic for ORDINARY inputs — it never overwrites an existing workspace,
  * preflights non-directory scaffold parents, and rolls back a partial write. It does not defend
  * against every pathological target state (TOCTOU, FIFOs, disk-full): recover by delete-and-retry.
+ *
+ * Sibling scaffold modules: add-channel.ts (`add <channel>`), vendor-skill.ts (`add skill`); the file
+ * content they and this module write lives in scaffold-templates.ts.
  */
-import { access, appendFile, cp, lstat, mkdir, readFile, readdir, rename, rm, writeFile } from "node:fs/promises";
-import { existsSync } from "node:fs";
-import { basename, dirname, isAbsolute, join, relative, resolve } from "node:path";
-import { homedir } from "node:os";
-import { type LoadedDefinition, assertInsideWorkspace, loadAgentDefinition, loadRootIgnore } from "./definition.ts";
+import { access, lstat, mkdir, readdir, rm, writeFile } from "node:fs/promises";
+import { dirname, join } from "node:path";
+import { loadRootIgnore } from "./definition.ts";
+import {
+  CONFIG_MJS,
+  ENV_EXAMPLE,
+  GITIGNORE,
+  SKILL_MD,
+  TOOL_TS,
+  agentsMd,
+  packageJson,
+  toPackageName,
+} from "./scaffold-templates.ts";
 import { fastagentVersion } from "./version.ts";
-
-/** Identity persona (clean — it is the system prompt). The complete variant references the tool. */
-function agentsMd(minimal: boolean): string {
-  const toolLine = minimal ? "" : "\nWhen the user asks how long a piece of text is, use the word-count tool.\n";
-  return `# Assistant
-
-You are a concise, helpful assistant. Answer directly and skip filler.
-
-When the user asks you to write or edit prose, consult the house-style skill first.
-${toolLine}`;
-}
-
-const SKILL_MD = `---
-name: house-style
-description: The house writing style. Consult before writing or editing any prose for the user.
----
-# House style
-
-- Prefer short sentences and the active voice.
-- Avoid marketing adjectives ("seamless", "powerful", "robust").
-- Lead with the answer; put caveats after.
-`;
-
-const TOOL_TS = `import { defineTool, z } from "@kid7st/fastagent";
-
-// A code tool: filename (word-count.ts) is the tool name. tools/ is auto-discovered,
-// so it needs no registration in fastagent.config. Test it directly with:
-//   fastagent tool word-count '{"text":"hello there world"}'
-export default defineTool({
-  description: "Count the words and characters in a piece of text.",
-  input: z.object({ text: z.string().describe("The text to measure") }),
-  async execute({ text }) {
-    const trimmed = text.trim();
-    return { words: trimmed ? trimmed.split(/\\s+/).length : 0, characters: text.length };
-  },
-});
-`;
-
-const CONFIG_MJS = `// fastagent.config.mjs — deployment choices only (model / http; code tools auto-discover from tools/).
-// Your agent's identity and behavior live in AGENTS.md + skills/ + tools/, never here.
-// Model precedence: \`--model\` flag > FASTAGENT_MODEL env > this default.
-// Change "model" to a "provider/modelId" you have access to (\`fastagent models\` lists them).
-export default {
-  model: "openai-codex/gpt-5.5",
-  http: { port: 8787 },
-};
-`;
-
-const GITIGNORE = `# secrets — never commit (kept out of git and any deploy copy)
-.env
-
-# dependencies (reinstalled at deploy)
-node_modules/
-
-# fastagent machine state (dev/start sessions)
-.fastagent/
-`;
-
-// All-commented: copying to .env sets nothing by accident, and every knob is optional. Auth leads
-// with `fastagent login` because the default model (openai-codex) is OAuth-only — never imply an API key.
-const ENV_EXAMPLE = `# Environment for this agent. Copy to .env (gitignored) and uncomment what you need.
-# Everything here is OPTIONAL — the defaults work without a .env.
-
-# --- Model auth ---
-# The default model (openai-codex) signs in with OAuth, not an API key: run \`fastagent login\` once.
-# Switch to an API-key provider? Set its key here — the variable name is provider-specific
-# (e.g. OPENAI_API_KEY, ANTHROPIC_API_KEY). Run \`fastagent models\` to see available specs.
-# OPENAI_API_KEY=
-# ANTHROPIC_API_KEY=
-
-# --- Model selection (overrides fastagent.config) ---
-# Precedence: --model flag > FASTAGENT_MODEL > config.
-# FASTAGENT_MODEL=openai-codex/gpt-5.5
-
-# --- Serving (fastagent start) ---
-# Port precedence: --port > PORT > config.http.port > 8787
-# PORT=8787
-# Where conversations persist (default: ./fastagent-sessions)
-# FASTAGENT_SESSIONS_DIR=./fastagent-sessions
-`;
-
-/** package.json for the complete agent: ESM + the deps a defineTool tool imports. The
- *  @kid7st/fastagent range tracks THIS build's version, so a fresh workspace installs an API-matching version. */
-function packageJson(name: string, version: string): string {
-  return `${JSON.stringify(
-    {
-      name,
-      private: true,
-      type: "module",
-      dependencies: { "@kid7st/fastagent": `^${version}`, zod: "^4.0.0" },
-    },
-    null,
-    2,
-  )}\n`;
-}
-
-/** Sanitize a directory basename into a valid npm package name (lowercase, safe chars). */
-function toPackageName(dir: string): string {
-  const base = basename(resolve(dir))
-    .toLowerCase()
-    .replace(/[^a-z0-9._-]/g, "-")
-    .replace(/^[._-]+/, "");
-  return base || "agent";
-}
 
 interface ScaffoldFile {
   rel: string;
@@ -143,375 +50,12 @@ export interface ScaffoldResult {
   warnings: string[];
 }
 
-async function exists(p: string): Promise<boolean> {
+/** Does a path exist? (async; shared with the sibling scaffold modules). */
+export async function exists(p: string): Promise<boolean> {
   return access(p).then(
     () => true,
     () => false,
   );
-}
-
-/** A scaffolded `channels/github.ts`: the third-party adapter import + a starter `on()` to edit. */
-const CHANNEL_GITHUB_TS = `import { githubChannel } from "@kid7st/fastagent/github";
-import type { ChannelModule } from "@kid7st/fastagent";
-
-// A channel = a third-party ADAPTER (githubChannel: verify + parse + ACK) wired to YOUR on() glue.
-// fastagent discovers this file under channels/ and serves the routes it returns. Set
-// GITHUB_WEBHOOK_SECRET in .env (a missing secret fails at startup — an empty key would accept forged
-// deliveries) and point a GitHub webhook (JSON) at POST /webhook.
-const channel: ChannelModule = (agent) => ({
-  "POST /webhook": githubChannel(agent, {
-    secret: process.env.GITHUB_WEBHOOK_SECRET ?? "",
-    // Map a verified event to the intents the agent acts on (empty array = ignore). Each review is
-    // INDEPENDENT and idempotent (it reconciles against the PR's existing comments), so use a
-    // distinct per-delivery session (event.deliveryId): overlapping deliveries then run on their
-    // own session without a shared-lease drop.
-    on: (event) => {
-      if (event.event === "pull_request" && event.action === "opened" && "pull_request" in event.payload) {
-        const { repository, pull_request } = event.payload;
-        return [
-          {
-            session: event.deliveryId,
-            text: \`Review pull request #\${pull_request.number} in \${repository.full_name}.\`,
-          },
-        ];
-      }
-      return [];
-    },
-  }),
-});
-
-export default channel;
-`;
-
-/** A scaffolded `channels/telegram.ts`: the Telegram adapter import + a starter `on()` to edit. */
-const CHANNEL_TELEGRAM_TS = `import { telegramChannel } from "@kid7st/fastagent/telegram";
-import type { ChannelModule } from "@kid7st/fastagent";
-
-// A channel = a third-party ADAPTER (telegramChannel: verify + run + reply) wired to YOUR on() glue.
-// fastagent discovers this file under channels/ and serves the routes it returns. Setup:
-//   1. @BotFather → /newbot → put the bot token in TELEGRAM_BOT_TOKEN
-//   2. pick a random TELEGRAM_SECRET_TOKEN (verifies that inbound updates really come from Telegram)
-//   3. register the webhook once, pointing Telegram at POST /telegram with that secret:
-//        curl "https://api.telegram.org/bot$TELEGRAM_BOT_TOKEN/setWebhook" \\
-//          -d url=https://your.host/telegram -d secret_token=$TELEGRAM_SECRET_TOKEN
-const channel: ChannelModule = (agent) => ({
-  "POST /telegram": telegramChannel(agent, {
-    secretToken: process.env.TELEGRAM_SECRET_TOKEN ?? "", // missing → fails at startup (would accept forged updates)
-    botToken: process.env.TELEGRAM_BOT_TOKEN ?? "",       // used to send the agent's reply back to the chat
-    // Dev/personal bot: surface raw errors to the chat so you (and your AI agent) can act on them. The
-    // chat is customer-facing by default — for a public bot, drop this or return a neutral string;
-    // full details always go to the server log regardless.
-    onError: (failed) => \`⚠️ \${failed.details}\`,
-    // The channel owns transport + format (HTML) + attachments (photo→vision, file→disk) + streaming.
-    // \`route\` (POLICY) is OPTIONAL — omitted, it uses defaultTelegramRoute: private chats always answer,
-    // groups only on a command / reply to the bot / @mention. Override to customise, reusing the export:
-    //   route: (u) => defaultTelegramRoute(u) && { session: \`user:\${u.message?.from?.id}\` },
-    //   route: (u) => defaultTelegramRoute(u) && { text: \`\${telegramEnvelope(u.message!)}\\n[extra]\` },
-  }),
-});
-
-export default channel;
-`;
-
-/** A scaffolded `tools/telegram-send.ts`: the outbound action — the agent uploads a local file to a chat. */
-const TELEGRAM_SEND_TOOL_TS = `import { defineTool, z } from "@kid7st/fastagent";
-import { readFile } from "node:fs/promises";
-import { basename } from "node:path";
-
-// Send a local file back to a Telegram chat. The agent passes a chatId it reads from the
-// [telegram: chat …] context line the channel injects. tools/ is auto-discovered — no registration.
-export default defineTool({
-  description:
-    "Send a local file to a Telegram chat (a document, or a photo if it is an image). Pass the chatId from the [telegram: chat …] context line.",
-  input: z.object({
-    chatId: z.union([z.string(), z.number()]).describe("target chat id (from the [telegram: chat …] context line)"),
-    path: z.string().describe("absolute path of the local file to send"),
-    caption: z.string().optional(),
-    asPhoto: z.boolean().optional().describe("send as a photo (inline) instead of a document"),
-    messageThreadId: z.number().optional().describe("thread to reply into (from the context line), if any"),
-  }),
-  async execute({ chatId, path, caption, asPhoto, messageThreadId }) {
-    const token = process.env.TELEGRAM_BOT_TOKEN;
-    if (!token) throw new Error("TELEGRAM_BOT_TOKEN is not set");
-    const method = asPhoto ? "sendPhoto" : "sendDocument";
-    const form = new FormData();
-    form.set("chat_id", String(chatId));
-    if (messageThreadId !== undefined) form.set("message_thread_id", String(messageThreadId));
-    if (caption) form.set("caption", caption);
-    form.set(asPhoto ? "photo" : "document", new Blob([await readFile(path)]), basename(path));
-    const res = await fetch("https://api.telegram.org/bot" + token + "/" + method, { method: "POST", body: form });
-    const data = (await res.json().catch(() => ({}))) as { ok?: boolean; description?: string };
-    if (!res.ok || !data.ok) throw new Error("telegram " + method + " failed: " + res.status + " " + (data.description ?? ""));
-    return "sent " + basename(path) + " to chat " + chatId;
-  },
-});
-`;
-
-export type ChannelKind = "github" | "telegram";
-
-/** An env var a scaffolded channel reads. `generate` = a random-string secret the CLI can pre-fill. */
-export interface ChannelEnv {
-  name: string;
-  hint: string;
-  generate?: boolean;
-}
-
-interface ChannelScaffold {
-  template: string;
-  env: ChannelEnv[];
-  /** Channel-specific next-step lines, printed after the env lines and before the `dev` line. */
-  steps: string[];
-  /** An optional companion tool dropped into `tools/` (e.g. an outbound action the agent calls). */
-  tool?: { name: string; template: string };
-}
-
-const CHANNEL_SCAFFOLDS: Record<ChannelKind, ChannelScaffold> = {
-  github: {
-    template: CHANNEL_GITHUB_TS,
-    env: [
-      {
-        name: "GITHUB_WEBHOOK_SECRET",
-        hint: "any random string; set the same value in the GitHub webhook",
-        generate: true,
-      },
-    ],
-    steps: [
-      "edit channels/github.ts — map events to intents in on()",
-      "add the webhook in your repo (Settings → Webhooks): Payload URL = <public-url>/webhook, content type application/json",
-    ],
-  },
-  telegram: {
-    template: CHANNEL_TELEGRAM_TS,
-    env: [
-      { name: "TELEGRAM_BOT_TOKEN", hint: "from @BotFather → /newbot" },
-      { name: "TELEGRAM_SECRET_TOKEN", hint: "any random string; verifies inbound updates", generate: true },
-    ],
-    steps: [
-      "edit channels/telegram.ts — customise routing with route() (optional; the defaults already work)",
-      "the agent can send files back by calling the scaffolded tools/telegram-send.ts tool",
-    ],
-    tool: { name: "telegram-send", template: TELEGRAM_SEND_TOOL_TS },
-  },
-};
-
-/** The channel kinds `fastagent add <kind>` can scaffold. */
-export const CHANNEL_KINDS = Object.keys(CHANNEL_SCAFFOLDS) as ChannelKind[];
-
-/** The env vars + next-step lines a scaffolded channel needs (for the CLI to print). */
-export function channelSetup(kind: ChannelKind): { env: ChannelEnv[]; steps: string[] } {
-  const { env, steps } = CHANNEL_SCAFFOLDS[kind];
-  return { env, steps };
-}
-
-/**
- * Append a channel's env vars (commented placeholders + hints) to `.env.example`, so a developer who
- * copies it to `.env` finds the vars already there. No-op when there is no `.env.example` or the block
- * is already present. Placeholders only — no real secret lands in the committable template.
- */
-export async function appendChannelEnv(dir: string, kind: ChannelKind): Promise<boolean> {
-  const file = join(dir, ".env.example");
-  let current: string;
-  try {
-    current = await readFile(file, "utf8");
-  } catch (e) {
-    if ((e as NodeJS.ErrnoException).code === "ENOENT") return false;
-    throw e;
-  }
-  const marker = `# --- ${kind} channel ---`;
-  if (current.includes(marker)) return false;
-  const block = `\n${marker}\n${CHANNEL_SCAFFOLDS[kind].env.map((e) => `# ${e.name}=   # ${e.hint}`).join("\n")}\n`;
-  await appendFile(file, block);
-  return true;
-}
-
-/** The path `add <kind>` scaffolds to. */
-function channelPath(dir: string, kind: ChannelKind): string {
-  return join(dir, "channels", `${kind}.ts`);
-}
-
-/** Whether a channel file already exists — checked before any mutation, so a no-clobber re-add is side-effect-free. */
-export async function channelExists(dir: string, kind: ChannelKind): Promise<boolean> {
-  return exists(channelPath(dir, kind));
-}
-
-/**
- * Scaffold `channels/<kind>.ts` into {@link dir}. Never clobbers an existing file (the `on()` glue is
- * authored content). The wx write is the TOCTOU safety net behind {@link channelExists}.
- */
-export async function scaffoldChannel(dir: string, kind: ChannelKind): Promise<string> {
-  const channelsDir = join(dir, "channels");
-  // Don't write through a channels/ symlink that escapes the workspace; an in-workspace one is fine.
-  await assertInsideWorkspace(dir, "channels");
-  const file = channelPath(dir, kind);
-  if (await exists(file)) {
-    throw new Error(`${file} already exists — edit it, or remove it to re-scaffold`);
-  }
-  await mkdir(channelsDir, { recursive: true });
-  await writeFile(file, CHANNEL_SCAFFOLDS[kind].template, { flag: "wx" });
-  // A companion tool (e.g. telegram's outbound send) goes in tools/; never clobber an authored one.
-  const tool = CHANNEL_SCAFFOLDS[kind].tool;
-  if (tool && !(await exists(join(dir, "tools", `${tool.name}.ts`)))) {
-    await mkdir(join(dir, "tools"), { recursive: true });
-    await writeFile(join(dir, "tools", `${tool.name}.ts`), tool.template, { flag: "wx" });
-  }
-  return file;
-}
-
-/**
- * Verify the workspace is ready to host a channel: an ESM package.json that declares
- * `@kid7st/fastagent` (the channel file imports it). `add` checks and guides, never bootstraps — that
- * is `init`'s job.
- */
-export async function assertChannelReady(dir: string): Promise<void> {
-  const pkgPath = join(dir, "package.json");
-  let raw: string;
-  try {
-    raw = await readFile(pkgPath, "utf8");
-  } catch (e) {
-    if ((e as NodeJS.ErrnoException).code === "ENOENT") {
-      throw new Error(`${dir}: not a fastagent code workspace (no package.json) — run \`fastagent init\` here first`);
-    }
-    throw e;
-  }
-  let pkg: { type?: string; dependencies?: Record<string, string> };
-  try {
-    pkg = JSON.parse(raw);
-  } catch {
-    throw new Error(`${pkgPath}: invalid JSON`);
-  }
-  if (pkg.type !== "module") {
-    throw new Error(`${pkgPath}: fastagent channels are ESM — set "type": "module"`);
-  }
-  if (typeof pkg.dependencies?.["@kid7st/fastagent"] !== "string") {
-    throw new Error(
-      `${pkgPath}: add "@kid7st/fastagent" to dependencies (then \`npm install\`) — the channel file imports it`,
-    );
-  }
-}
-
-/** Derive the destination skill name from a source ref: the last path segment, sans `#ref`. */
-function skillNameFromSource(source: string): string {
-  const noRef = source.split("#")[0] ?? source;
-  return basename(noRef.replace(/\/+$/, ""));
-}
-
-/** A local source is an explicit path (./x, ../x, /abs); anything else is a giget ref or a bare name. */
-function isLocalSource(source: string): boolean {
-  return source.startsWith("./") || source.startsWith("../") || source === "." || isAbsolute(source);
-}
-
-/** A bare name (no `/`, no `scheme:`) resolves against the local global skill dirs (see below). */
-function isBareName(source: string): boolean {
-  return !source.includes("/") && !/^[a-z][a-z0-9+.-]*:/i.test(source);
-}
-
-/** Local "global" skill dirs, used ONLY as an add-time vendoring source (a bare `add skill <name>`
- *  copies the match in, git-tracked) — nothing is loaded from here at run time. */
-function findGlobalSkillSource(name: string): string | undefined {
-  for (const root of [join(homedir(), ".agents", "skills"), join(homedir(), ".pi", "agent", "skills")]) {
-    if (existsSync(join(root, name, "SKILL.md"))) return join(root, name);
-  }
-  return undefined;
-}
-
-export interface VendoredSkill {
-  /** The skill's real name (from SKILL.md frontmatter, per the Agent Skills spec). */
-  name: string;
-  description?: string;
-  /** Workspace-relative destination (e.g. `skills/pdf`). */
-  dest: string;
-  /** The skill ships a `scripts/` dir (executable code) — a trust signal for the caller. */
-  hasScripts: boolean;
-  /** Spec diagnostics for THIS skill from the runtime loader (e.g. name ≠ dir). */
-  diagnostics: LoadedDefinition["diagnostics"];
-  /** True when an existing skill was overwritten (--update); false for a fresh vendor. */
-  overwritten: boolean;
-}
-
-/**
- * Vendor an Agent Skills skill into `<workspace>/skills/<name>/` from a giget ref (github default), a
- * local path, or a bare name (resolved against the local global skill dirs). Copy-in, git-tracked.
- * Refuses to overwrite unless `options.update` (then a plain git-tracked overwrite, never a merge).
- * Validates a staging copy with the runtime loader BEFORE replacing, so a bad fetch never destroys an
- * existing skill.
- */
-export async function vendorSkill(
-  workspaceDir: string,
-  source: string,
-  options: { update?: boolean } = {},
-): Promise<VendoredSkill> {
-  const name = skillNameFromSource(source);
-  if (name === "" || name === "." || name === "..") {
-    throw new Error(`cannot derive a skill name from "${source}" — point at a skill directory (…/skills/<name>)`);
-  }
-  const skillsDir = join(workspaceDir, "skills");
-  const dest = join(skillsDir, name);
-  // Refuse to clobber unless --update; the check is side-effect-free, and a git-tracked overwrite is
-  // safe (review with `git diff`, undo with `git checkout`).
-  const overwritten = existsSync(dest);
-  if (overwritten && !options.update) {
-    throw new Error(
-      `skills/${name} already exists — re-run with --update to overwrite it (git tracks the change), or remove it`,
-    );
-  }
-  await mkdir(skillsDir, { recursive: true });
-
-  // Fetch into a STAGING dir (same filesystem → atomic rename), validate, and only THEN replace dest,
-  // so a failed/invalid fetch never destroys an existing skill. The leading "." keeps the loader from
-  // treating staging as a skill.
-  const staging = join(skillsDir, `.${name}.vendoring`);
-  await rm(staging, { recursive: true, force: true }); // clear any leftover from a prior crash
-  try {
-    if (isLocalSource(source)) {
-      const src = resolve(source);
-      if (!existsSync(join(src, "SKILL.md"))) {
-        throw new Error(`"${source}" has no SKILL.md — an Agent Skills skill is a directory containing SKILL.md`);
-      }
-      await cp(src, staging, { recursive: true });
-    } else if (isBareName(source)) {
-      // bare name → vendor from a local global skill dir (add-time copy, not a runtime scan).
-      const src = findGlobalSkillSource(source);
-      if (!src) {
-        throw new Error(
-          `no skill "${source}" in your global skill dirs (~/.agents/skills, ~/.pi/agent/skills) — ` +
-            `give a git ref (owner/repo/path) or a local path instead`,
-        );
-      }
-      await cp(src, staging, { recursive: true });
-    } else {
-      // giget defaults a BARE ref to its own template registry, not github — so default the provider to
-      // github for a plain `owner/repo/path` (an explicit `github:`/`gh:`/`gitlab:`… scheme is kept).
-      // Supports a subdir + #ref, fetched via the tar API (no git binary).
-      const ref = /^[a-z][a-z0-9+.-]*:/i.test(source) ? source : `github:${source}`;
-      // Lazy import: giget is only needed for a git ref, so the serve path (index.ts → init.ts) and the
-      // local/bare-name sources never load it.
-      const { downloadTemplate } = await import("giget");
-      await downloadTemplate(ref, { dir: staging, force: true });
-    }
-    if (!existsSync(join(staging, "SKILL.md"))) {
-      throw new Error(`"${source}" did not yield a SKILL.md — expected an Agent Skills skill directory`);
-    }
-  } catch (error) {
-    await rm(staging, { recursive: true, force: true }); // failed/invalid: drop staging, leave dest intact
-    throw error;
-  }
-  // Validated. Replace atomically — the old skill survived every failure path above.
-  if (overwritten) await rm(dest, { recursive: true, force: true });
-  await rename(staging, dest);
-
-  // Report via the runtime loader, matching THIS skill by EXACT directory (a substring match would
-  // prefix-pollute a sibling `<name>-x` and break on Windows path separators).
-  const def = await loadAgentDefinition(workspaceDir);
-  const rel = join("skills", name);
-  const skill = def.skills.find((sk) => relative(workspaceDir, dirname(sk.filePath)) === rel);
-  return {
-    name: skill?.name ?? name,
-    description: skill?.description,
-    dest: rel,
-    hasScripts: existsSync(join(dest, "scripts")),
-    diagnostics: def.diagnostics.filter((d) => d.path !== undefined && relative(workspaceDir, dirname(d.path)) === rel),
-    overwritten,
-  };
 }
 
 /**
